added script to produce semantic similarity graph
added compare_texts.py, a simple script that takes in a directory of
documents and uses the tf-idf cosine similarity metric to produce an
upper triangular numpy matrix representing the edge weights of this
graph. This matrix can then be used to produce and plot a networkx graph
if networkx is installed.