added tuple to dict converter for key/value pairs
Since the tokenizer returns nested tuples, when we access elements,
it's never very clear what element is being accessed since they are
indexed with numbers instead of strings.

KeyValueToDict is a class that allows us to convert these nested tuples
into dictionaries so that they can be indexed using strings, which
helps with readability in the reducers.